Fix negative latency issue on Windows (issue #40)
This hunk is the main fix for the issue:
	-       return (Time.QuadPart * 1000000000I64 / Frequency.QuadPart);
	+       return ((Time.QuadPart * 1000000)/Frequency.QuadPart);

+++ src/lagscope.h
@@ -66,8 +66,8 @@ struct lagscope_test_client{
 
 struct lagscope_test_runtime{
 	struct lagscope_test *test;
-	long long start_time;
-	long long current_time;
+	double start_time;
+	double current_time;
 	unsigned long  ping_elapsed;
 
 	unsigned long lazy_prog_report_factor;
